apps/api/pyproject.toml
[project]
name = "liderix-api"
version = "0.1.0"
description = "Backend API for Liderix Platform"
authors = [
    { name = "Kirill", email = "kprolieiev@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
    "uvicorn[standard]>=0.34.3,<0.35.0",
    "fastapi>=0.115.14,<0.116.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "alembic>=1.16.2,<2.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "pydantic[email]>=2.11.7,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-jose[cryptography]>=3.5.0,<4.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "greenlet>=3.2.3,<4.0.0",
    "bcrypt==4.0.1",
    "python-dotenv>=1.0.1,<2.0.0"  # 🆕 для работы с .env
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.2"